refactored chat history and reused codes
