+++ src/solace_ai_connector/components/general/litellm/litellm_chat_model_with_history.py
@@ -3,6 +3,7 @@
 import time
 
 from .litellm_base import LiteLLMChatModelBase, litellm_info_base
+from ..utils.chat_history_handler import ChatHistoryHandler
 from ....common.log import log
 
 info = litellm_info_base.copy()
@@ -32,7 +33,7 @@ info["input_schema"]["properties"]["clear_history_but_keep_depth"] = {
 }
 
 
-class LiteLLMChatModelWithHistory(LiteLLMChatModelBase):
+class LiteLLMChatModelWithHistory(LiteLLMChatModelBase, ChatHistoryHandler):
     def __init__(self, **kwargs):
         super().__init__(info, **kwargs)
         self.history_max_turns = self.get_config("history_max_turns", 10)
@@ -52,12 +53,12 @@ class LiteLLMChatModelWithHistory(LiteLLMChatModelBase):
             if clear_history_but_keep_depth is not None:
                 clear_history_but_keep_depth = max(0, int(clear_history_but_keep_depth))
         except (TypeError, ValueError):
+            log.error("Invalid clear_history_but_keep_depth value. Defaulting to 0.")
             clear_history_but_keep_depth = 0
         messages = data.get("messages", [])
 
         with self.get_lock(self.history_key):
             history = self.kv_store_get(self.history_key) or {}
-
             if session_id not in history:
                 history[session_id] = {"messages": [], "last_accessed": time.time()}
 
@@ -67,6 +68,7 @@ class LiteLLMChatModelWithHistory(LiteLLMChatModelBase):
                 )
 
             session_history = history[session_id]["messages"]
+            log.debug(f"Session history: {session_history}")
 
             # If the passed in messages have a system message and the history's
             # first message is a system message, then replace the history's first
@@ -99,73 +101,6 @@ class LiteLLMChatModelWithHistory(LiteLLMChatModelBase):
             )
 
             self.kv_store_set(self.history_key, history)
+            log.debug(f"Updated history: {history}")
 
         return response
-
-    def prune_history(self, session_id, history):
-        current_time = time.time()
-        if current_time - history[session_id]["last_accessed"] > self.history_max_time:
-            history[session_id]["messages"] = []
-        elif len(history[session_id]["messages"]) > self.history_max_turns * 2:
-            history[session_id]["messages"] = history[session_id]["messages"][
-                -self.history_max_turns * 2 :
-            ]
-        self.make_history_start_with_user_message(session_id, history)
-
-    def clear_history_but_keep_depth(self, session_id: str, depth: int, history):
-        if session_id in history:
-            messages = history[session_id]["messages"]
-            # If the depth is 0, then clear all history
-            if depth == 0:
-                history[session_id]["messages"] = []
-                history[session_id]["last_accessed"] = time.time()
-                return
-
-            # Check if the history is already shorter than the depth
-            if len(messages) <= depth:
-                # Do nothing, since the history is already shorter than the depth
-                return
-
-            # If the message at depth is not a user message, then
-            # increment the depth until a user message is found
-            while depth < len(messages) and messages[-depth]["role"] != "user":
-                depth += 1
-            history[session_id]["messages"] = messages[-depth:]
-            history[session_id]["last_accessed"] = time.time()
-
-            # In the unlikely case that the history starts with a non-user message,
-            # remove it
-            self.make_history_start_with_user_message(session_id, history)
-
-    def make_history_start_with_user_message(self, session_id, history):
-        if session_id in history:
-            messages = history[session_id]["messages"]
-            if messages:
-                if messages[0]["role"] == "system":
-                    # Start from the second message if the first is "system"
-                    start_index = 1
-                else:
-                    # Start from the first message otherwise
-                    start_index = 0
-
-                while (
-                    start_index < len(messages)
-                    and messages[start_index]["role"] != "user"
-                ):
-                    messages.pop(start_index)
-
-    def handle_timer_event(self, timer_data):
-        if timer_data["timer_id"] == "history_cleanup":
-            self.history_age_out()
-
-    def history_age_out(self):
-        with self.get_lock(self.history_key):
-            history = self.kv_store_get(self.history_key) or {}
-            current_time = time.time()
-            for session_id in list(history.keys()):
-                if (
-                    current_time - history[session_id]["last_accessed"]
-                    > self.history_max_time
-                ):
-                    del history[session_id]
-            self.kv_store_set(self.history_key, history)
