+++ src/solace_ai_connector/components/general/openai/openai_chat_model_with_history.py
@@ -4,6 +4,7 @@ import time
 
 import time
 from .openai_chat_model_base import OpenAIChatModelBase, openai_info_base
+from ..utils.chat_history_handler import ChatHistoryHandler
 
 info = openai_info_base.copy()
 info["class_name"] = "OpenAIChatModelWithHistory"
@@ -33,7 +34,7 @@ info["input_schema"]["properties"]["clear_history_but_keep_depth"] = {
 }
 
 
-class OpenAIChatModelWithHistory(OpenAIChatModelBase):
+class OpenAIChatModelWithHistory(OpenAIChatModelBase, ChatHistoryHandler):
     def __init__(self, **kwargs):
         super().__init__(info, **kwargs)
         self.history_max_turns = self.get_config("history_max_turns", 10)
@@ -102,71 +103,3 @@ class OpenAIChatModelWithHistory(OpenAIChatModelBase):
             self.kv_store_set(self.history_key, history)
 
         return response
-
-    def prune_history(self, session_id, history):
-        current_time = time.time()
-        if current_time - history[session_id]["last_accessed"] > self.history_max_time:
-            history[session_id]["messages"] = []
-        elif len(history[session_id]["messages"]) > self.history_max_turns * 2:
-            history[session_id]["messages"] = history[session_id]["messages"][
-                -self.history_max_turns * 2 :
-            ]
-        self.make_history_start_with_user_message(session_id, history)
-
-    def clear_history_but_keep_depth(self, session_id: str, depth: int, history):
-        if session_id in history:
-            messages = history[session_id]["messages"]
-            # If the depth is 0, then clear all history
-            if depth == 0:
-                history[session_id]["messages"] = []
-                history[session_id]["last_accessed"] = time.time()
-                return
-
-            # Check if the history is already shorter than the depth
-            if len(messages) <= depth:
-                # Do nothing, since the history is already shorter than the depth
-                return
-
-            # If the message at depth is not a user message, then
-            # increment the depth until a user message is found
-            while depth < len(messages) and messages[-depth]["role"] != "user":
-                depth += 1
-            history[session_id]["messages"] = messages[-depth:]
-            history[session_id]["last_accessed"] = time.time()
-
-            # In the unlikely case that the history starts with a non-user message,
-            # remove it
-            self.make_history_start_with_user_message(session_id, history)
-
-    def make_history_start_with_user_message(self, session_id, history):
-        if session_id in history:
-            messages = history[session_id]["messages"]
-            if messages:
-                if messages[0]["role"] == "system":
-                    # Start from the second message if the first is "system"
-                    start_index = 1
-                else:
-                    # Start from the first message otherwise
-                    start_index = 0
-
-                while (
-                    start_index < len(messages)
-                    and messages[start_index]["role"] != "user"
-                ):
-                    messages.pop(start_index)
-
-    def handle_timer_event(self, timer_data):
-        if timer_data["timer_id"] == "history_cleanup":
-            self.history_age_out()
-
-    def history_age_out(self):
-        with self.get_lock(self.history_key):
-            history = self.kv_store_get(self.history_key) or {}
-            current_time = time.time()
-            for session_id in list(history.keys()):
-                if (
-                    current_time - history[session_id]["last_accessed"]
-                    > self.history_max_time
-                ):
-                    del history[session_id]
-            self.kv_store_set(self.history_key, history)
